+++ src/containers/modals/MultiStepWithdrawalModal/config.ts
@@ -13,8 +13,7 @@ export const steps = [
   },
   {
     label: 'Prove Withdrawal',
-    description:
-      'You will be able to submit your proof once your transaction is submitted to L1.',
+    description: 'You can submit your proof once your transaction reaches L1.',
     passiveStep: false,
     btnLbl: 'Prove Withdrawal',
   },
